Find and mark commits squashable into some earlier commit
Squashable in this case means:
 * that there are no collisions between the commits, and
 * that the lines that changes in the commit are a subset of
   the lines that changes in the earlier commit.

The idea is that if you have made small later changes to new code, it might
simplify the commit history to squash them into the earlier commit. By
including only changes that are subsets of an earlier commit in the
fragmap, the resulting changes will not destroy some important
intermediate state of code that was pre-existing, which might be undesirable.